refactor(sections): update imports to shared
